+++ packages/snap/snap.manifest.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/blockchain-lab-um/masca.git"
   },
   "source": {
-    "shasum": "bQ0/zCgnnYwBhEs477ZyQC0D01R9Dj6bkaSeAN0wNrg=",
+    "shasum": "xGtrmIyF3k39+muQz0YvaXkHBhpL6FhIJ01ndqen0/0=",
     "location": {
       "npm": {
         "filePath": "dist/snap.js",
