+++ packages/snap/src/rpc/oidc/handleOIDCCredentialOffer.ts
@@ -1,29 +1,23 @@
-import type {
-  AvailableVCStores,
-  HandleOIDCCredentialOfferRequestParams,
-} from '@blockchain-lab-um/masca-types';
-import type {
-  SendOIDCAuthorizationResponseArgs,
-  SignArgs,
-} from '@blockchain-lab-um/oidc-client-plugin';
+import type { HandleOIDCCredentialOfferRequestParams } from '@blockchain-lab-um/masca-types';
+import type { SignArgs } from '@blockchain-lab-um/oidc-client-plugin';
 import type {
   CredentialRequest,
   TokenResponse,
 } from '@blockchain-lab-um/oidc-types';
 import { isError, Result } from '@blockchain-lab-um/utils';
 import { heading, panel } from '@metamask/snaps-ui';
-import type { UnsignedPresentation, VerifiableCredential } from '@veramo/core';
+import type { VerifiableCredential } from '@veramo/core';
 import { decodeCredentialToObject } from '@veramo/utils';
-import qs from 'qs';
 
 import type { ApiParams } from '../../interfaces';
 import { getCurrentDid } from '../../utils/didUtils';
 import { snapGetKeysFromAddress } from '../../utils/keyPair';
-import { sign } from '../../utils/sign';
 import {
-  veramoImportMetaMaskAccount,
-  veramoQueryVCs,
-} from '../../utils/veramoUtils';
+  handleAuthorizationRequest,
+  sendAuthorizationResponse,
+} from '../../utils/oidc';
+import { sign } from '../../utils/sign';
+import { veramoImportMetaMaskAccount } from '../../utils/veramoUtils';
 import { getAgent } from '../../veramo/setup';
 
 export async function handleOIDCCredentialOffer(
@@ -44,6 +38,9 @@ export async function handleOIDCCredentialOffer(
     bip44CoinTypeNode,
   });
 
+  if (did.startsWith('did:ethr') || did.startsWith('did:pkh'))
+    throw new Error('did:ethr and did:pkh are not supported');
+
   const agent = await getAgent(snap, ethereum);
 
   const credentialOfferResult = await agent.parseOIDCCredentialOfferURI({
@@ -54,33 +51,7 @@ export async function handleOIDCCredentialOffer(
     throw new Error(credentialOfferResult.error);
   }
 
-  const credentialOffer = credentialOfferResult.data;
-
-  const { credentials, grants } = credentialOffer;
-
-  const isPinRequired =
-    grants?.['urn:ietf:params:oauth:grant-type:pre-authorized_code']
-      ?.user_pin_required ?? false;
-
-  let pin;
-
-  // Ask user for PIN
-  if (isPinRequired) {
-    pin = await snap.request({
-      method: 'snap_dialog',
-      params: {
-        type: 'prompt',
-        content: panel([
-          heading('Please enter the PIN you received from the issuer'),
-        ]),
-        placeholder: 'PIN...',
-      },
-    });
-
-    if (!pin || typeof pin !== 'string') {
-      throw new Error('PIN is required');
-    }
-  }
+  const { credentials, grants } = credentialOfferResult.data;
 
   const res = await snapGetKeysFromAddress({
     snap,
@@ -91,13 +62,24 @@ export async function handleOIDCCredentialOffer(
 
   if (res === null) throw new Error('Could not get keys from address');
 
-  const kid = did.startsWith('did:ethr')
-    ? `${did}#controllerKey`
-    : `${did}#${did.split(':')[2]}`;
+  await veramoImportMetaMaskAccount(
+    {
+      snap,
+      ethereum,
+      state,
+      account,
+      bip44CoinTypeNode,
+    },
+    agent
+  );
+
+  // TODO: Is this fine or should we improve it ?
+  const kid = `${did}#${did.split(':')[2]}`;
 
   const isDidKeyEbsi =
     state.accountState[account].accountConfig.ssi.didMethod === 'did:key:ebsi';
 
+  // TODO: Select curve based on the key in the identifier ?
   const customSign = async (args: SignArgs) =>
     sign(args, {
       privateKey: res.privateKey,
@@ -121,164 +103,58 @@ export async function handleOIDCCredentialOffer(
 
     console.log(`authorizationRequestURI: ${authorizationRequestURI}`);
 
-    const authorizationRequestResult =
-      await agent.parseOIDCAuthorizationRequestURI({
-        authorizationRequestURI,
-      });
+    const handleAuthorizationRequestResult = await handleAuthorizationRequest({
+      agent,
+      authorizationRequestURI,
+      did,
+      customSign,
+    });
 
-    if (isError(authorizationRequestResult)) {
-      throw new Error(authorizationRequestResult.error);
-    }
-
-    const authorizationRequest = authorizationRequestResult.data;
-    const sendOIDCAuthorizationResponseArgs: SendOIDCAuthorizationResponseArgs =
-      {};
-
-    if (authorizationRequest.response_type.includes('id_token')) {
-      // Create id token
-      const idTokenResult = await agent.createIdToken({
-        sign: customSign,
-      });
-
-      if (isError(idTokenResult)) {
-        throw new Error(idTokenResult.error);
-      }
-
-      const idToken = idTokenResult.data;
-
-      sendOIDCAuthorizationResponseArgs.idToken = idToken;
-    }
-
-    if (authorizationRequest.response_type.includes('vp_token')) {
-      const store = ['snap'] as AvailableVCStores[];
-
-      const queryResults = await veramoQueryVCs({
-        snap,
-        ethereum,
-        options: { store, returnStore: false },
-      });
-
-      const queriedCredentials: any = queryResults.map((result) => result.data);
-
-      console.log('queriedCredentials');
-      console.log(queriedCredentials);
-
-      const selectCredentialsResult = await agent.selectCredentials({
-        credentials: queriedCredentials,
-      });
-
-      console.log('selectCredentialsResult');
-      console.log(selectCredentialsResult);
-
-      if (isError(selectCredentialsResult)) {
-        throw new Error(selectCredentialsResult.error);
-      }
-
-      const createPresentationSubmissionResult =
-        await agent.createPresentationSubmission({
-          credentials: selectCredentialsResult.data,
-        });
-
-      if (isError(createPresentationSubmissionResult)) {
-        throw new Error(createPresentationSubmissionResult.error);
-      }
-
-      const presentationSubmission = createPresentationSubmissionResult.data;
-
-      const decodedCredentials = selectCredentialsResult.data.map(
-        // eslint-disable-next-line @typescript-eslint/no-unsafe-return
-        (credential) => decodeCredentialToObject(credential).proof.jwt
+    if (handleAuthorizationRequestResult.isUserInteractionRequired) {
+      throw new Error(
+        'User interaction is required. This is not supported yet'
       );
-
-      const identifier = await veramoImportMetaMaskAccount(
-        {
-          snap,
-          ethereum,
-          state,
-          account,
-          bip44CoinTypeNode,
-        },
-        agent
-      );
-
-      // const getChallengeResult = await agent.getChallenge();
-
-      // if (isError(getChallengeResult)) {
-      //   throw new Error(getChallengeResult.error);
-      // }
-
-      // const challenge = getChallengeResult.data;
-
-      // const getDomainResult = await agent.getDomain();
-
-      // if (isError(getDomainResult)) {
-      //   throw new Error(getDomainResult.error);
-      // }
-
-      // const domain = getDomainResult.data;
-
-      const veramoPresentation = await agent.createVerifiablePresentation({
-        presentation: {
-          holder: did,
-          verifiableCredential: decodedCredentials,
-        },
-        proofFormat: 'jwt',
-      });
-
-      const { '@context': context, holder, type } = veramoPresentation;
-
-      const vp: UnsignedPresentation = {
-        '@context': context,
-        holder,
-        type,
-        verifiableCredential: decodedCredentials,
-      };
-
-      const createVpTokenResult = await agent.createVpToken({
-        sign: customSign,
-        vp,
-      });
-
-      if (isError(createVpTokenResult)) {
-        throw new Error(createVpTokenResult.error);
-      }
-
-      const vpToken = createVpTokenResult.data;
-
-      sendOIDCAuthorizationResponseArgs.presentationSubmission =
-        presentationSubmission;
-      sendOIDCAuthorizationResponseArgs.vpToken = vpToken;
     }
 
-    // POST /auth-mock/direct_post
-    const authorizationResponseResult =
-      await agent.sendOIDCAuthorizationResponse(
-        sendOIDCAuthorizationResponseArgs
-      );
+    const { sendOIDCAuthorizationResponseArgs } =
+      handleAuthorizationRequestResult;
 
-    if (isError(authorizationResponseResult)) {
-      throw new Error(authorizationResponseResult.error);
-    }
+    const sendAuthorizationResponseResult = await sendAuthorizationResponse({
+      agent,
+      sendOIDCAuthorizationResponseArgs,
+    });
 
-    const authorizationResponse = authorizationResponseResult.data;
-
-    const authorizationResponseData: any = qs.parse(
-      authorizationResponse.split('?')[1]
-    );
-
-    if (!authorizationResponseData.code) {
-      throw new Error('Authorization code is required');
-    }
-
-    if (!authorizationResponseData.state) {
-      throw new Error('State is required');
-    }
+    const { code } = sendAuthorizationResponseResult;
 
     tokenRequestResult = await agent.sendTokenRequest({
-      code: authorizationResponseData.code,
+      code,
       clientId: did,
     });
   } else if (grants?.['urn:ietf:params:oauth:grant-type:pre-authorized_code']) {
+    // Check if PIN is required
+    const isPinRequired =
+      grants['urn:ietf:params:oauth:grant-type:pre-authorized_code']
+        .user_pin_required ?? false;
+
+    let pin;
+
+    // Ask user for PIN
+    if (isPinRequired) {
+      pin = await snap.request({
+        method: 'snap_dialog',
+        params: {
+          type: 'prompt',
+          content: panel([
+            heading('Please enter the PIN you received from the issuer'),
+          ]),
+          placeholder: 'PIN...',
+        },
+      });
+
+      if (!pin || typeof pin !== 'string') {
+        throw new Error('PIN is required');
+      }
+    }
     tokenRequestResult = await agent.sendTokenRequest(pin ? { pin } : {});
   } else {
     throw new Error('Unsupported grant type');
@@ -326,8 +202,6 @@ export async function handleOIDCCredentialOffer(
   }
 
   credentialRequest.proof = proofOfPossessionResult.data;
-
-  // if(did.startsWith('did:ethr') || did.startsWith('did:pkh')) throw new Error('did:ethr and did:pkh are not supported');
 
   const credentialRequestResult = await agent.sendCredentialRequest(
     credentialRequest
